+++ inst/shiny/shark-qc/global.R
@@ -7,25 +7,6 @@ source("helpers.R")
 clean_shark4r_cache(days = 1, verbose = FALSE)
 
 # Translate datatype names
-type_lookup <- c(
-  "Bacterioplankton"       = "Bacterioplankton",
-  "Chlorophyll"            = "Chlorophyll",
-  "Epibenthos"             = "Epibenthos",
-  "Grey seal"              = "GreySeal",
-  "Harbour Porpoise"       = "HarbourPorpoise",
-  "Harbour seal"           = "HarbourSeal",
-  "Physical and Chemical"  = "PhysicalChemical",
-  "Phytoplankton"          = "Phytoplankton",
-  "Picoplankton"           = "Picoplankton",
-  "Plankton Barcoding"     = "PlanktonBarcoding",
-  "Plankton Imaging"       = "PlanktonImaging",
-  "Primary production"     = "PrimaryProduction",
-  "Profile"                = "Profile",
-  "Ringed seal"            = "RingedSeal",
-  "Seal pathology"         = "SealPathology",
-  "Sedimentation"          = "Sedimentation",
-  "Zoobenthos"             = "Zoobenthos",
-  "Zooplankton"            = "Zooplankton"
-)
+type_lookup <- SHARK4R:::.type_lookup
 
 shark_codes <- get_shark_codes()