+++ NEWS.md
@@ -10,6 +10,7 @@
 * Add new function `clean_shark4r_cache()` to clear cached files.
 * Add new function `check_setup()` and `run_qc_app()` to download and run SHARK QC scripts and Shiny App.
 * Add new function `get_shark_statistics()` to download SHARK data and extract summary statistics for numeric parameters.
+* Add new function `translate_shark_datatype()` to translate user-facing datatype names to internal SHARK4R names.
 
 ## Enhancements
 * Re-export functions from the `iRfcb` package: `which_basin()` and `positions_are_near_land()`.